requirements.txt
Flask==2.3.3
msal>=1.30.0
python-dotenv==1.0.0
requests==2.31.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Session==0.5.0
gunicorn==21.2.0  # For production server
psycopg2-binary==2.9.9  # For PostgreSQL support in Azure 